// apps/web/src/modules/screens/definitions.generated.ts
export interface ScreenDef {
  title: string
  columns: string[]
  fields: FormFieldDef[]
}